refactor: user_ component를 => user component로 명칭 변경
