+++ src/models/index.ts
@@ -1,5 +1,4 @@
 import userDao from "./user.dao";
-import user_Dao from "./user_.dao";
 import postFormDao from "./postForm.dao";
 import cmtDao from "./comment.dao";
 import categoryDao from "./category.dao";
@@ -11,7 +10,6 @@ import postDao from "./post.dao";
 
 const dao_set = {
     userDao,
-    user_Dao,
     postFormDao,
     cmtDao,
     categoryDao,
