+++ src/services/index.ts
@@ -1,6 +1,5 @@
 import cmtSvc from "./comment.service";
 import userSvc from "./user.service";
-import user_Svc from "./user_.service";
 import cateSvc from "./category.service";
 import postFormSvc from "./postForm.service";
 import postSvc from "./post.service";
@@ -12,7 +11,6 @@ const svc_set = {
     cmtSvc,
     postSvc,
     userSvc,
-    user_Svc,
     memberRequestSvc,
     companyRequestSvc,
     cateSvc,
