+++ src/services/postForm.service.ts
@@ -1,6 +1,6 @@
 import daoset from '../models';
 
-const { postFormDao, companyDao, user_Dao, categoryDao, locationDao } = daoset;
+const { postFormDao, companyDao, userDao, categoryDao, locationDao } = daoset;
 
 const putPostForm = async (postFormInput: CompanyPostFormInput) => {
   let result;
@@ -22,7 +22,7 @@ const putPostForm = async (postFormInput: CompanyPostFormInput) => {
   }
 
   // 유저 정보 조회
-  const userInfo = await user_Dao.findUser({userId: Number(postFormInput.usersId)});
+  const userInfo = await userDao.findUser({userId: Number(postFormInput.usersId)});
   console.log("userInfo: ", userInfo);
   if (!userInfo) {
     throw {status: 404, message: "유저 ID에 해당하는 유저 정보가 존재하지 않습니다."};
